refactor(i18n): fold the content-language helpers in with the locale ones
`content_language.ts` was a module for two one-line functions. They belong
beside `normalizeLocale` and `getFormattingLocale` in `formatters.ts`, which
despite its name already holds locale-derived behaviour rather than only
formatting — `getMeasurementSystem` picks miles over kilometres by exactly the
same kind of lookup.

`i18n.ts` would have been the obvious home and cannot be: `server.ts` imports
`t` from it, so giving it an `options` import would close an
`i18n -> options -> server -> i18n` cycle. `formatters.ts` already imports
both, so this adds no edge to the graph.

Its spec mocks `../services/i18n` wholesale, so the mock factory now supplies
`getLocaleById` from a dynamic import of the shared catalog rather than a stub,
which could otherwise disagree with the locales the assertions name.
