+++ apps/client/src/utils/formatters.spec.ts
@@ -3,14 +3,20 @@ import { describe, expect, it, vi } from "vitest";
 // i18next is not initialised here, so have t() echo the key and the interpolated count. That lets the
 // tests assert exactly which unit was chosen and what count was computed — the real logic — while
 // leaving the plural rendering itself to i18next.
-vi.mock("../services/i18n", () => ({
-    t: (key: string, opts?: { count?: number }) => `${key}|${opts?.count}`
-}));
+// `getLocaleById` is the real lookup rather than a stub — it is a find over the shared catalog, so
+// mocking it would only risk disagreeing with the locales the assertions name.
+vi.mock("../services/i18n", async () => {
+    const { LOCALES } = await import("@triliumnext/commons");
+    return {
+        t: (key: string, opts?: { count?: number }) => `${key}|${opts?.count}`,
+        getLocaleById: (id: string | null | undefined) => LOCALES.find((l) => l.id === id) ?? null
+    };
+});
 
 import { LOCALES } from "@triliumnext/commons";
 
 import options from "../services/options";
-import { formatDateNumeric, formatDateTime, formatDuration, getMeasurementSystem, normalizeLocale } from "./formatters";
+import { formatDateNumeric, formatDateTime, formatDuration, getMeasurementSystem, isContentRightToLeft, normalizeLocale, resolveContentLanguage } from "./formatters";
 
 describe("formatters", () => {
     it("tolerates incorrect locale", () => {
@@ -254,5 +260,49 @@ describe("formatters", () => {
             // 7 is not one of the offered scales, so the unit cannot be named from it.
             expect(formatDuration(604800, 7)).toBe("time_interval.days|7");
         });
+    });
+});
+
+describe("resolveContentLanguage", () => {
+    it("prefers the note's own language over the configured default", () => {
+        options.set("defaultContentLanguage", "fr");
+        options.set("locale", "ru");
+
+        expect(resolveContentLanguage("de")).toBe("de");
+    });
+
+    it("falls back to the default content language, then to the application's language", () => {
+        options.set("defaultContentLanguage", "fr");
+        options.set("locale", "ru");
+
+        expect(resolveContentLanguage(null)).toBe("fr");
+        expect(resolveContentLanguage(undefined)).toBe("fr");
+        expect(resolveContentLanguage("")).toBe("fr");
+
+        // An empty default is the "auto" entry, meaning follow the application's language rather
+        // than meaning no language at all.
+        options.set("defaultContentLanguage", "");
+        expect(resolveContentLanguage(null)).toBe("ru");
+    });
+});
+
+describe("isContentRightToLeft", () => {
+    it("follows the note's own language", () => {
+        expect(isContentRightToLeft("he")).toBe(true);
+        expect(isContentRightToLeft("de")).toBe(false);
+    });
+
+    it("applies the default to a note that has no language of its own", () => {
+        options.set("defaultContentLanguage", "ar");
+        expect(isContentRightToLeft(null)).toBe(true);
+
+        // ...and a note that does have one is not dragged along by it.
+        expect(isContentRightToLeft("de")).toBe(false);
+    });
+
+    it("treats an unrecognized language as left-to-right", () => {
+        options.set("defaultContentLanguage", "not-a-locale");
+        expect(isContentRightToLeft(null)).toBe(false);
+        expect(isContentRightToLeft("also-not-a-locale")).toBe(false);
     });
 });
